Updated namespaces and plugin paths in NivoSlider
Updated the log message in `WidgetsNivoSliderViewComponent.cs` to reflect the new namespace `SIDS.NewProductsSlider`. Also, the path to the `Configure.cshtml` view in `WidgetsNivoSliderController.cs` has been updated to point to the `SIDS.NewProductsSlider` plugin. These changes are due to a renaming or restructuring of the project's namespaces and plugins.