Update naming convention, target framework, and supported versions
Updated the `AreaNames.Admin` constant to `AreaNames.ADMIN` in `WidgetsNivoSliderController.cs` for consistency or to address a case sensitivity issue. Upgraded the `TargetFramework` from `net7.0` to `net8.0` in `SIDS.Plugin.Misc.NewProductsSlider.csproj`, indicating a shift to the .NET 8.0 framework. Also, updated the `SupportedVersions` from `4.60` to `4.70` in `plugin.json`, showing that the plugin now supports version 4.70.