Limit slider to 5 products, remove catalog number check
The call to `LimitCatalogNumberRepitition` on `nivoSliderProducts` has been commented out, so the function will no longer be applied to the list of products. Additionally, the `foreach` loop that iterates over `nivoSliderProducts` has been modified to only take the first 5 products from the list using the `Take(5)` method.

These changes mean the code no longer limits the catalog number repetition for the products marked as new, and the display of slider pictures is now limited to a maximum of 5 products.